requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-google-genai>=1.0.0
tavily-python>=0.3.0
duckduckgo-search>=6.0.0
tweepy>=4.14.0
linkedin-api>=2.0.0
beautifulsoup4>=4.12.0
playwright>=1.44.0
pydantic>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
langsmith>=0.1.0
aiohttp>=3.9.0
langchain-community>=0.2.0
json-repair>=0.25.0
requests>=2.31.0
